Added XRPD mineral quantifications
Mineral quantification (based on XY files) for a subset of about 500 samples. Note that the XRDP data was merged with another subset of the AfSIS dataset (radiocarbon data). Those two datasets do not always match exactly, so sometimes the closest XRPD sample was used to match with the other AfSIS dataset (thus, there are two SSN ID columns that do not always match).